+++ CHANGES.rst
@@ -10,6 +10,8 @@ Added
 - Display exit code in parentheses after error message.
 - Do not reformat renamed files.
 - CI workflow to post recent project activity in a discussion. Triggered manually.
+- The ``--preview`` configuration flag is now supported in the configuration files for
+  Darker and Black
 
 Fixed
 -----